+++ src/config.js
@@ -122,6 +122,13 @@ export default Object.assign(
      */
     KEEP_SESSION_INTERVAL: 1000 * 60 * 2,
 
+    /**
+     * Change over different recovery modes. To disable recovery, leave the
+     * field empty or give it any unacceptable value, i.e.: 'none'.
+     *
+     * Acceptable values: 'seed' - enables recovery by seed
+     */
+    RECOVERY_MODE: 'seed',
   },
   // process.env,
   process.env