+++ src/vuex/types.js
@@ -10,11 +10,12 @@ const mutations = {
   SET_KV_ENTRY_BLOCKED_ROLE_ID: 'SET_KV_ENTRY_BLOCKED_ROLE_ID',
   SET_KV_ENTRY_US_VERIFIED_ROLE_ID: 'SET_KV_ENTRY_US_VERIFIED_ROLE_ID',
   SET_KV_ENTRY_US_ACCREDITED_ROLE_ID: 'SET_KV_ENTRY_US_ACCREDITED_ROLE_ID',
+  SET_KV_ASSET_TYPE_DEFAULT: 'SET_KV_ASSET_TYPE_DEFAULT',
+  SET_KV_ASSET_TYPE_KYC_REQUIRED: 'SET_KV_ASSET_TYPE_KYC_REQUIRED',
   SET_KV_ASSET_TYPE_SECURITY: 'SET_KV_ASSET_TYPE_SECURITY',
   SET_KV_POLL_TYPE_RESTRICTED: 'SET_KV_POLL_TYPE_RESTRICTED',
   SET_KV_POLL_TYPE_UNRESTRICTED: 'SET_KV_POLL_TYPE_UNRESTRICTED',
 
-  SET_KV_KYC_REQUIRED: 'SET_KV_KYC_REQUIRED',
   SET_DEFAULT_QUOTE_ASSET: 'SET_DEFAULT_QUOTE_ASSET',
 
   // account
@@ -91,6 +92,7 @@ const getters = {
   kvEntryBlockedRoleId: 'kvEntryBlockedRoleId',
   kvEntryUsVerifiedRoleId: 'kvEntryUsVerifiedRoleId',
   kvEntryUsAccreditedRoleId: 'kvEntryUsAccreditedRoleId',
+  kvAssetTypeDefault: 'kvAssetTypeDefault',
   kvAssetTypeKycRequired: 'kvAssetTypeKycRequired',
   kvAssetTypeSecurity: 'kvAssetTypeSecurity',
   kvPollTypeRestricted: 'kvPollTypeRestricted',
@@ -147,7 +149,9 @@ const getters = {
   // assets
   assets: 'assets',
   assetByCode: 'assetByCode',
+  assetsByOwner: 'assetsByOwner',
   balancesAssets: 'balancesAssets',
+  balancesAssetsByOwner: 'balancesAssetsByOwner',
   fiatAssets: 'fiatAssets',
   depositableAssets: 'depositableAssets',
   coinpaymentsAssets: 'coinpaymentsAssets',
@@ -155,8 +159,9 @@ const getters = {
   withdrawableBalancesAssets: 'withdrawableBalancesAssets',
   statsQuoteAsset: 'statsQuoteAsset',
   ownedAssets: 'ownedAssets',
+  ownedBalancesAssets: 'ownedBalancesAssets',
   baseAtomicSwapBalancesAssets: 'baseAtomicSwapBalancesAssets',
-  quoteAtomicSwapBalancesAssets: 'quoteAtomicSwapBalancesAssets',
+  quoteAtomicSwapAssets: 'quoteAtomicSwapAssets',
 }
 
 export const vuexTypes = {