+++ CHANGELOG.md
@@ -8,11 +8,36 @@ Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
 ## [Unreleased]
+#### Fixed
+- Resolved an issue when in select field quote assets on the atomic swap form
+ show only assets in user balance
+- Fixed display price on form when create new sale
+
+## [1.10.0-x.0] - 2019-07-02
 #### Added
 - New Atomic swap feature:
   - Create atomic swap form
 - `canBeBaseInAtomicSwap` and `canBeQuoteInAtomicSwap` policy for assets
 - Integration with session key server
+
+### "Under the hood" changes
+#### Changed
+- Now using @tokend/js-sdk@1.9.0-x.0
+
+### Experimental features changes
+#### Added
+- New use case - conto
+
+## [1.9.0-rc.1] - 2019-07-08
+#### Fixed
+- Resolved an issue when has absent margin and present label overlap on
+  limits changing form
+- Resolved an issue when you can set amount bigger than max asset’s amount
+- Resolved an issue when limits leftover amount was not updated
+- Resolved an issue when poll failed to load for a user that is not sale owner
+- Resolved an issue when participant emails did not show on the "Participants"
+  tab
+- Resolved an issue when the user could not update a sale request
 
 ## [1.9.0-rc.0] - 2019-06-24
 #### Added
@@ -808,7 +833,9 @@ for further information about branching and tagging conventions.
 
 ## [1.3.0] - 2019-03-01
 
-[Unreleased]: https://github.com/tokend/web-client/compare/1.9.0-rc.0...HEAD
+[Unreleased]: https://github.com/tokend/web-client/compare/1.10.0-x.0...HEAD
+[1.10.0-x.0]: https://github.com/tokend/web-client/compare/1.9.0-rc.1...1.10.0-x.0
+[1.9.0-rc.1]: https://github.com/tokend/web-client/compare/1.9.0-rc.0...1.9.0-rc.1
 [1.9.0-rc.0]: https://github.com/tokend/web-client/compare/1.8.0...1.9.0-rc.0
 [1.8.0]: https://github.com/tokend/web-client/compare/1.8.0-rc.3...1.8.0
 [1.8.0-rc.3]: https://github.com/tokend/web-client/compare/1.8.0-rc.2...1.8.0-rc.3